// src/components/types/types.all.ts
export interface IDetalhesPaciente {
    address: [] | any
    agendaConsultas: [] | any
    alergiaMedicamento: string
    beneficioSocial: string
    caps: boolean
    cid: string
    cns: string
    condutaImediata: string
    conjuge: string
    consultas: [] | any
    cor: string
    cpf: string
    createAt: Date
    doenca: string
    drogas: string
    escolaridade: string
    familiaVuneravel: string
    frequencia: string
    genero: string
    id: string
    inicioTratamento: string
    mae: string
    motivoAcolhimento: string
    nascimento: Date
    nome: string
    pai: string
    prontuario: string
    servicoEncaminhado: string
    tecResponsavel: string
    telContato: string
    usaMedicacao: string
}